pyproject.toml
[tool.poetry]
name = "car-list"
version = "0.1.0"
description = ""
authors = ["Madiyar Tasbolatov <aizhir.mt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
